Factor out the timing control logic into it's own model, and implement an actual FSM for it.
Unfortunately while the cleanup was nice, this added about 26 more LUTs to the design for only
a marginal speed improvement.  It seems that switching the default value for t_next to t+1 and
dropping the final "else" clause on all of the states drops the LUT count back down to something
reasonable but I wanted to capture this design.