Add 'next' outputs for data out and write signals so an external block ram could be used with only a single port.
This meant having to rework instructions that do writes to have the write flag set a clock earlier in the microcode than it was before.  The external behavior for the asynchronous address/read/etc signals doesn't change, though.

I also had to add support for inhibiting the write in the no-carry case for some of the indexed addressing modes.   I also went back to having dedicated ALU support for the TRB and RMB instructions since I needed to do things in a different order.