Work in progress on Mega65 style hypervisor.
I can actually jump into it now with the current implementation.  I'm not yet
sure if I'm going to stay on the current path or not but I may finish the current
implementation.   I have another idea that would do it more with custom microcode
in the CPU that might require less logic overall and wouldn't require a custom
external state machine that has to interpose the CPU path to memory.