More hypervisor work.  Fleshed out registers and trap mechanism a bit more.
I'm still not happy with how many external resources this takes.  I may
try yet a third approach next which would be to add a 68000 style supervisor
stack pointer or something so that a hypervisor trap could be done more like
a 68000 trap where we push user state (PC and P) onto the supervisor stack.